requirements.txt
altair==5.1.2
attrs==23.1.0
blinker==1.7.0
boto3==1.34.74
botocore==1.34.74
cachetools==5.3.2
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
Cython==3.0.5
et-xmlfile==1.1.0
gitdb==4.0.11
GitPython==3.1.40
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
mpmath==1.3.0
mysql-connector-python==8.2.0
numpy==1.26.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.3
patsy==0.5.3
Pillow==10.1.0
plotly==5.18.0
pmdarima==2.0.4
protobuf==4.21.12
psycopg2-binary==2.9.9
pyarrow==14.0.1
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
pytz==2023.3.post1
referencing==0.30.2
requests==2.31.0
rich==13.6.0
rpds-py==0.12.0
s3transfer==0.10.1
scikit-learn==1.2.2
scipy==1.11.3
setuptools==69.1.1
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.23
statsmodels==0.14.0
streamlit==1.28.2
sympy==1.12
tenacity==8.2.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==1.25.4
validators==0.22.0
zipp==3.17.0
